feat: update package.json to specify Node.js engine version
fix: rename getBookmarked method to getBookmarkeds in BookmarkController

refactor: format checkout routes for better readability

feat: add scheduled command for expiring subscriptions in console routes

feat: implement webhook route for handling Xendit invoice notifications

chore: add Vercel configuration file for deployment

+++ database/migrations/2025_04_18_025652_create_orders_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->decimal('total_amount', 10, 2)->nullable();
             $table->decimal('discount_amount', 10, 2)->default(0);
             $table->decimal('final_amount', 10, 2)->nullable();
-            $table->bigInteger('total_admin_fee')->default(0);
+            $table->decimal('total_admin_fee', 10, 2)->default(0);
             $table->string('payment_method', 50)->nullable();
             $table->string('payment_provider', 50)->nullable();
             $table->string('status')->default('pending');
